@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:clima/utilities/constants.dart';
+import '../utilities/constants.dart';
 
 class CityScreen extends StatefulWidget {
   @override
@@ -23,7 +23,7 @@ class _CityScreenState extends State<CityScreen> {
             children: <Widget>[
               Align(
                 alignment: Alignment.topLeft,
-                child: FlatButton(
+                child: TextButton(
                   onPressed: () {},
                   child: Icon(
                     Icons.arrow_back_ios,
@@ -35,7 +35,7 @@ class _CityScreenState extends State<CityScreen> {
                 padding: EdgeInsets.all(20.0),
                 child: null,
               ),
-              FlatButton(
+              TextButton(
                 onPressed: () {},
                 child: Text(
                   'Get Weather',
