@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:clima/utilities/constants.dart';
+import '../utilities/constants.dart';
 
 class LocationScreen extends StatefulWidget {
   @override
@@ -28,14 +28,14 @@ class _LocationScreenState extends State<LocationScreen> {
               Row(
                 mainAxisAlignment: MainAxisAlignment.spaceBetween,
                 children: <Widget>[
-                  FlatButton(
+                  TextButton(
                     onPressed: () {},
                     child: Icon(
                       Icons.near_me,
                       size: 50.0,
                     ),
                   ),
-                  FlatButton(
+                  TextButton(
                     onPressed: () {},
                     child: Icon(
                       Icons.location_city,
